+++ bb-rsi/src/main/java/com/ticker/bbrsi/rx/BbRsiThread.java
@@ -25,11 +25,11 @@ public class BbRsiThread extends StratThread<BbRsiService> {
 
     private long triggerWaveEndTime;
     private long tradeStartTime;
-    private float tradeValue;
-    private float targetValue = 0;
-    private float peak;
-    private float dip;
-    private float rsiPrev;
+    private double tradeValue;
+    private double targetValue = 0;
+    private double peak;
+    private double dip;
+    private double rsiPrev;
     private long rsiSetTime;
     private int panicSell = 0;
     private int panicBuy = 0;
@@ -46,7 +46,7 @@ public class BbRsiThread extends StratThread<BbRsiService> {
     }
 
     @Override
-    public void setRsi(float rsi) {
+    public void setRsi(double rsi) {
         if (!service.isSameMinTrigger(rsiSetTime)) {
             rsiPrev = rsi;
             rsiSetTime = System.currentTimeMillis();
@@ -59,7 +59,7 @@ public class BbRsiThread extends StratThread<BbRsiService> {
      *
      * @return the rsi diff
      */
-    public float getRsiDiff() {
+    public double getRsiDiff() {
         return getRsi() - getRsiPrev();
     }
 
@@ -68,7 +68,7 @@ public class BbRsiThread extends StratThread<BbRsiService> {
      *
      * @param peak the peak
      */
-    public void setPeak(float peak) {
+    public void setPeak(double peak) {
         if (this.peak == 0 || this.peak < peak) {
             log.debug(getThreadName() + " : peak changed from " + this.peak);
             this.peak = peak;
@@ -81,7 +81,7 @@ public class BbRsiThread extends StratThread<BbRsiService> {
      *
      * @param dip the dip
      */
-    public void setDip(float dip) {
+    public void setDip(double dip) {
         if (this.dip == 0 || this.dip > dip) {
             log.debug(getThreadName() + " : dip changed from " + this.dip);
             this.dip = dip;
